+++ Assets/Project/Scripts/Contexts/DamageContext.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 public class DamageContext {
     public Character Attacker;
     public Character Target;
-    public int BaseDamage;    // 計算前ダメージ
-    public int FinalDamage;   // 計算後ダメージ
+    public float BaseDamage;    // 計算前ダメージ
+    public float FinalDamage;   // 計算後ダメージ
     public bool IsCritical;
 }
